Tracks done (albeit OCDifying)
All tracks are done in the sense that everything has found a routing, some components still may need to be moved though, such as the sensors on board and the push switches. 
Antenna stuff needs doing.
Board may need to be reshaped (keepout zone?).
GND plane needs to be added + silkscreens + memes